requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==22.2.0
autobahn==24.4.2
Automat==24.8.1
avro-python3==1.10.2
certifi==2022.12.7
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==2.1.1
confluent-kafka==2.1.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.4
elastic-apm==6.15.1
fastavro==1.7.3
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.8
multidict==6.0.4
mysqlclient==2.2.4
openai==0.27.0
packaging==22.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.0.2
pyOpenSSL==24.2.1
python-dateutil==2.8.2
python-logstash==0.4.8
pytz==2022.6
redis==5.0.8
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.15.0
yarl==1.8.2
zope.interface==7.0.1
